feat: enable Amplify DataStore with test frontend
This commit enables the Amplify DataStore feature. A Goal schema was
added to test this. After the API for this schema was deployed to the
backend, I added a basic form to App.js that can be used to create
goals and list them out to verify that it is written to the backend.

No security or account features have been implemented at this time.
It is anticipated that most of this will be rewritten to use React
Native list components on different pages. The backend may be changed
to use AWS RDS rather than AWS DynamoDB, which is used by Amplify.

small steps eh.

JIRA: TYMORA-29